requirements.txt
torch==1.0.1
torchvision
pretrainedmodels
numpy
opencv-python-headless
joblib
albumentations
scikit-image
tqdm
glog
tensorboardx
fire
pyyaml
# this file is not ready yet
